added SettingsControlDialog
