+++ renderer/src/components/SelectProgram/SelectProgram.jsx
@@ -5,34 +5,23 @@ import PropTypes from "prop-types";
 import React, { useCallback, useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { useNavigate } from "react-router-dom";
-import {
-  deleteProgram,
-  useGetBootQuery,
-  useGetProgramsQuery,
-  useGetSettingsQuery,
-} from "../../api/ipc";
-import { ERRORS } from "../../constants/commonConst";
+import { deleteProgram, useGetProgramsQuery } from "../../api/ipc";
 import { PAGES, PAGES_PATHS, SUB_PATHS } from "../../constants/pathConst";
-import { BOOT_CONFIG_OPT, FILE_CONST } from "../../constants/reduxConst";
 import { SP_MODE } from "../../constants/selectProgramConst";
-import { MIN_MOTOR_STROKE, MOTOR_FIELDS } from "../../constants/settingsConst";
 import {
   TRANSLATION_KEYS,
   TRANSLATION_ROOT_KEYS,
 } from "../../constants/translationConst";
 import { getTranslationPath } from "../../utils/translationUtils";
 import DialogCustom from "../DialogCustom/DialogCustom";
-import ErrorText from "../ErrorText/ErrorText";
-import { EngineMotorElectroIcon } from "../Icons";
+import SettingsControlDialog from "../SettingsControlDialog/SettingsControlDialog";
 import { Container, Item } from "../SquareGrid/SquareGrid";
 
-const { SELECT_PROGRAM, PROGRAM_EDITOR, SETTINGS } = PAGES;
-const { COMMON_TRK, PROGRAM_EDITOR_TRK, TIPS_TRK } = TRANSLATION_ROOT_KEYS;
-const { deleteTKey, cancelTKey, copyTKey, errorTKey } =
-  TRANSLATION_KEYS[COMMON_TRK];
+const { SELECT_PROGRAM, PROGRAM_EDITOR } = PAGES;
+const { COMMON_TRK, PROGRAM_EDITOR_TRK } = TRANSLATION_ROOT_KEYS;
+const { deleteTKey, cancelTKey, copyTKey } = TRANSLATION_KEYS[COMMON_TRK];
 const { deleteProgHead, deleteProgMsg, copyProgHead, copyProgMsg } =
   TRANSLATION_KEYS[PROGRAM_EDITOR_TRK];
-const { motorBut } = TRANSLATION_KEYS[TIPS_TRK];
 
 const TARGET_DEFAULT = null;
 
@@ -45,40 +34,6 @@ const SelectProgram = props => {
     useGetProgramsQuery(undefined, {
       refetchOnMountOrArgChange: true,
     }) || {};
-  const { data: settings = {} } =
-    useGetSettingsQuery(undefined, {
-      refetchOnMountOrArgChange: true,
-      skip: mode !== SP_MODE.SELECT,
-    }) || {};
-  const minPosition = get(
-    settings,
-    [FILE_CONST.PERIPHERAL, MOTOR_FIELDS.MIN_POS],
-    null,
-  );
-  const maxPosition = get(
-    settings,
-    [FILE_CONST.PERIPHERAL, MOTOR_FIELDS.MAX_POS],
-    null,
-  );
-  const { data: bootConfig } =
-    useGetBootQuery(undefined, { refetchOnMountOrArgChange: true }) || {};
-  const spiStatus = get(bootConfig, [BOOT_CONFIG_OPT.SPI], null);
-  const isValidSettings =
-    isFinite(minPosition) &&
-    isFinite(maxPosition) &&
-    minPosition < maxPosition &&
-    Math.abs(minPosition - maxPosition) > MIN_MOTOR_STROKE &&
-    spiStatus;
-
-  const goToSettingsMotor = useCallback(() => {
-    navigate(
-      `${PAGES_PATHS[SETTINGS]}/${SUB_PATHS[SETTINGS].PERIPHERAL}/${SUB_PATHS[SETTINGS].MOTOR}`,
-    );
-  }, [navigate]);
-
-  const goToSettingsSpi = useCallback(() => {
-    navigate(`${PAGES_PATHS[SETTINGS]}/${SUB_PATHS[SETTINGS].PERFORMANCE}`);
-  }, [navigate]);
 
   const onDialogClose = useCallback(() => {
     switch (mode) {
@@ -220,41 +175,7 @@ const SelectProgram = props => {
         </Container>
       ))}
 
-      <DialogCustom
-        isOpen={mode === SP_MODE.SELECT && !isValidSettings}
-        icon={IconNames.WARNING_SIGN}
-        title={t(getTranslationPath(COMMON_TRK, errorTKey))}
-        canOutsideClickClose={false}
-        isCloseButtonShown={false}
-        body={
-          <ErrorText
-            error={
-              !spiStatus
-                ? ERRORS.BOOT_CONFIG_SPI_OFF
-                : ERRORS.INVALID_MOTOR_SETTINGS
-            }
-          />
-        }
-        footerMinimal
-        goBackBtn
-        footer={
-          !spiStatus ? (
-            <Button
-              large
-              icon={IconNames.COG}
-              text="TODO SPI"
-              onClick={goToSettingsSpi}
-            />
-          ) : (
-            <Button
-              large
-              icon={<EngineMotorElectroIcon />}
-              text={t(getTranslationPath(TIPS_TRK, motorBut))}
-              onClick={goToSettingsMotor}
-            />
-          )
-        }
-      />
+      {mode === SP_MODE.SELECT && <SettingsControlDialog />}
 
       <DialogCustom
         isOpen={Boolean(target)}
